[FIX] Cast embeddings to float32 before FAISS search to prevent segfault

+++ byokg-rag/src/graphrag_toolkit/byokg_rag/indexing/dense_index.py
@@ -68,7 +68,7 @@ class LocalFaissDenseIndex(DenseIndex):
         :return:
         """
 
-        query_emb = np.array(self.embedding.embed(input)).reshape(1, -1)
+        query_emb = np.array(self.embedding.embed(input), dtype=np.float32).reshape(1, -1)
 
         if id_selector is None:
             D, I = self.faiss_index.search(query_emb, topk)
@@ -94,7 +94,7 @@ class LocalFaissDenseIndex(DenseIndex):
         if id_selector is not None:
             raise NotImplementedError(f"id_selector not implemented for {self.__class__.__name__}")
 
-        query_emb = np.array(self.embedding.batch_embed(inputs)).reshape(len(inputs), -1)
+        query_emb = np.array(self.embedding.batch_embed(inputs), dtype=np.float32).reshape(len(inputs), -1)
 
         D, I = self.faiss_index.search(query_emb, topk)
 
@@ -133,9 +133,9 @@ class LocalFaissDenseIndex(DenseIndex):
             None
         """
         if isinstance(embeddings, np.ndarray):
-            doc_embs = embeddings
+            doc_embs = embeddings.astype(np.float32)
         else:
-            doc_embs = np.array(self.embedding.batch_embed(documents))
+            doc_embs = np.array(self.embedding.batch_embed(documents), dtype=np.float32)
         self.doc_store.extend(documents)
         self.doc_ids.extend(ids)
         for _id in ids:
